+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- `calcite-notice` now has a `text-label-close` attribute that allows the title of the close button to be set. It defaults to the English "close".
+
 ### Breaking Changes
 
 - `calcite-label` - `calciteLabelSelectedEvent` event has been renamed to `calciteLabelFocus`